+++ MonoGame.Randomchaos.Services.P2P/Enums/ConnectionTypesEnum.cs
@@ -1,10 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace SampleMonoGame.Randomchaos.Services.P2P.Enums
+﻿namespace MonoGame.Randomchaos.Services.P2P.Enums
 {
     public enum ConnectionTypesEnum
     {
