+++ MonoGame.Randomchaos.Services.P2P/Enums/ProtocolTypesEnum.cs
@@ -1,4 +1,10 @@
-﻿namespace SampleMonoGame.Randomchaos.Services.P2P.Enums
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace MonoGame.Randomchaos.Services.P2P.Enums
 {
     public enum ProtocolTypesEnum : int
     {
