+++ MonoGame.Randomchaos.Services.P2P/Interfaces/IClientPacketData.cs
@@ -1,18 +1,16 @@
-﻿
-using SampleMonoGame.Randomchaos.Services.P2P.Enums;
-using SampleMonoGame.Randomchaos.Services.P2P.Interfaces;
+﻿using MonoGame.Randomchaos.Services.P2P.Enums;
 using System;
 using System.Net;
 
-namespace SampleMonoGame.Randomchaos.Services.P2P.Models
+namespace MonoGame.Randomchaos.Services.P2P.Interfaces
 {
     ///-------------------------------------------------------------------------------------------------
-    /// <summary>   A client packet data. </summary>
+    /// <summary>   Interface for client packet data. </summary>
     ///
     /// <remarks>   Charles Humphrey, 17/10/2023. </remarks>
     ///-------------------------------------------------------------------------------------------------
 
-    public class ClientPacketData : IClientPacketData
+    public interface IClientPacketData
     {
         ///-------------------------------------------------------------------------------------------------
         /// <summary>   Gets or sets the identifier. </summary>
@@ -20,7 +18,7 @@ namespace SampleMonoGame.Randomchaos.Services.P2P.Models
         /// <value> The identifier. </value>
         ///-------------------------------------------------------------------------------------------------
 
-        public Guid Id { get; set; } = Guid.NewGuid();
+        Guid Id { get; set; }
 
         ///-------------------------------------------------------------------------------------------------
         /// <summary>   Gets or sets the UDP address. </summary>
@@ -28,7 +26,7 @@ namespace SampleMonoGame.Randomchaos.Services.P2P.Models
         /// <value> The UDP address. </value>
         ///-------------------------------------------------------------------------------------------------
 
-        public string UdpAddress { get; set; }
+        string UdpAddress { get; set; }
 
         ///-------------------------------------------------------------------------------------------------
         /// <summary>   Gets or sets the ud p port. </summary>
@@ -36,7 +34,7 @@ namespace SampleMonoGame.Randomchaos.Services.P2P.Models
         /// <value> The ud p port. </value>
         ///-------------------------------------------------------------------------------------------------
 
-        public int UdPPort { get; set; }
+        int UdPPort { get; set; }
 
         ///-------------------------------------------------------------------------------------------------
         /// <summary>   Gets or sets the type of the connection. </summary>
@@ -44,7 +42,7 @@ namespace SampleMonoGame.Randomchaos.Services.P2P.Models
         /// <value> The type of the connection. </value>
         ///-------------------------------------------------------------------------------------------------
 
-        public ConnectionTypesEnum ConnectionType { get; set; }
+        ConnectionTypesEnum ConnectionType { get; set; }
 
         ///-------------------------------------------------------------------------------------------------
         /// <summary>   Gets or sets information describing the player game. </summary>
@@ -52,35 +50,14 @@ namespace SampleMonoGame.Randomchaos.Services.P2P.Models
         /// <value> Information describing the player game. </value>
         ///-------------------------------------------------------------------------------------------------
 
-        public object? PlayerGameData { get; set; }
-
-        ///-------------------------------------------------------------------------------------------------
-        /// <summary>   Returns a string that represents the current object. </summary>
-        ///
-        /// <remarks>   Charles Humphrey, 17/10/2023. </remarks>
-        ///
-        /// <returns>   A string that represents the current object. </returns>
-        ///-------------------------------------------------------------------------------------------------
-
-        public override string ToString()
-        {
-            if (!string.IsNullOrEmpty(UdpAddress))
-                return $"{UdpAddress}:{UdPPort}";
-            else
-                return $"UDP Endpoint Unknown";
-        }
+        object? PlayerGameData { get; set; }
 
         ///-------------------------------------------------------------------------------------------------
         /// <summary>   Gets IP end point. </summary>
         ///
-        /// <remarks>   Charles Humphrey, 17/10/2023. </remarks>
-        ///
         /// <returns>   The IP end point. </returns>
         ///-------------------------------------------------------------------------------------------------
 
-        public virtual IPEndPoint GetIPEndPoint()
-        {
-            return new IPEndPoint(IPAddress.Parse(UdpAddress), UdPPort);
-        }
+        IPEndPoint GetIPEndPoint();
     }
 }
